refactor: StatusPage 이름 변경 (StatePage)
