+++ frontend/src/components/common/SideBar/SideBar.js
@@ -17,7 +17,7 @@ import {
   LogoutIcon,
   ReviewIcon,
   MyPageIcon,
-  StatusIcon,
+  StateIcon,
 } from '../../../assets/icons';
 import { useSnackBar } from '../../../hooks';
 
@@ -64,8 +64,8 @@ const SideBar = () => {
         <NavLinkElement to={PATH.REVIEW} activeStyle={selectedNavStyles}>
           <ReviewIcon width="20" height="20" stroke="currentColor" /> 접종후기
         </NavLinkElement>
-        <NavLinkElement to={PATH.STATUS} activeStyle={selectedNavStyles}>
-          <StatusIcon width="20" height="20" stroke="currentColor" /> 접종현황
+        <NavLinkElement to={PATH.STATE} activeStyle={selectedNavStyles}>
+          <StateIcon width="20" height="20" stroke="currentColor" /> 접종현황
         </NavLinkElement>
         {!!Object.keys(user).length && (
           <NavLinkElement
