+++ frontend/src/components/common/SideBarMobile/SideBarMobile.js
@@ -26,7 +26,7 @@ import {
   ShotVerificationMenuIcon,
   LikeIcon,
   CommentIcon,
-  StatusIcon,
+  StateIcon,
 } from '../../../assets/icons';
 import Button from '../Button/Button';
 import { css } from '@emotion/react';
@@ -96,8 +96,8 @@ const SideBarMobile = ({ isOpenSideBar, setIsOpenSideBar }) => {
           <NavLinkElement to={PATH.REVIEW} onClick={() => setIsOpenSideBar(false)}>
             <ReviewIcon width="23" height="23" stroke="currentColor" /> 접종후기
           </NavLinkElement>
-          <NavLinkElement to={PATH.STATUS} onClick={() => setIsOpenSideBar(false)}>
-            <StatusIcon width="23" height="23" stroke="currentColor" /> 접종현황
+          <NavLinkElement to={PATH.STATE} onClick={() => setIsOpenSideBar(false)}>
+            <StateIcon width="23" height="23" stroke="currentColor" /> 접종현황
           </NavLinkElement>
           {!!Object.keys(user).length && (
             <MyPageMenuContainer>
