+++ frontend/src/pages/index.js
@@ -11,7 +11,7 @@ import MyPageShotVerification from './MyPageShotVerification/MyPageShotVerificat
 import MyPageReview from './MyPageReview/MyPageReview';
 import MyPageCommentReview from './MyPageCommentReview/MyPageCommentReview';
 import MyPageLikeReview from './MyPageLikeReview/MyPageLikeReview';
-import StatusPage from './StatusPage/StatusPage';
+import StatePage from './StatePage/StatePage';
 
 export {
   HomePage,
@@ -27,5 +27,5 @@ export {
   MyPageReview,
   MyPageCommentReview,
   MyPageLikeReview,
-  StatusPage,
+  StatePage,
 };
